+++ src/MediaLibraryServiceProvider.php
@@ -7,7 +7,6 @@ use Spatie\MediaLibrary\Commands\CleanCommand;
 use Spatie\MediaLibrary\Commands\ClearCommand;
 use Spatie\MediaLibrary\Filesystem\Filesystem;
 use Spatie\MediaLibrary\Commands\RegenerateCommand;
-use Spatie\MediaLibrary\Filesystem\DefaultFilesystem;
 use Spatie\MediaLibrary\ResponsiveImages\WidthCalculator\WidthCalculator;
 use Spatie\MediaLibrary\ResponsiveImages\TinyPlaceholderGenerator\TinyPlaceholderGenerator;
 
@@ -50,7 +49,7 @@ class MediaLibraryServiceProvider extends ServiceProvider
         $this->app->bind('command.medialibrary:clear', ClearCommand::class);
         $this->app->bind('command.medialibrary:clean', CleanCommand::class);
 
-        $this->app->bind(Filesystem::class, DefaultFilesystem::class);
+        $this->app->bind(Filesystem::class, Filesystem::class);
 
         $this->app->bind(WidthCalculator::class, config('medialibrary.responsive_images.width_calculator'));
         $this->app->bind(TinyPlaceholderGenerator::class, config('medialibrary.responsive_images.tiny_placeholder_generator'));